Trying out some additional optimizations:
    - Removed the interface for grid, now using directly the templated
      class. It should theoritically improve performance by avoiding polymorphism
    - Callbacks in grid_travseral should now be inlined
    - Moved grid_traversal to algorithms, needs some splitting.

    I might aim for a header only library
